Split keyusage on its own function
